work started on gmime-2.5.x
2009-04-24  Jeffrey Stedfast  <fejj@novell.com>

	* build/vs2008/config-win32.h.in: New template to replace
	config.h.win32 so that the version info gets autogenerated when I
	make a new release. Makes it easier for me to keep it in sync.

	* build/vs2008/gmime.vcproj: No longer defines the gmime version
	variables on the command-line (as these are now part of config.h).

2009-04-23  Jeffrey Stedfast  <fejj@novell.com>

	* branched: now begins 2.5.x development

	* configure.in: Updated.

	* zentimer.h: Updated to hopefully build on Windows.

	* gmime/Makefile.am: Rename lib name to gmime-2.6.

	* gmime.pc.in: Renamed from gmime-2.4.pc.in.

	* mono/gmime-sharp.pc.in: Renamed from gmime-sharp-2.4.pc.in.

	* gmime/gmime-stream*.[c,h] (stream_length): Fixed to return a gint64

+++ gmime/gmime-stream-cat.c
@@ -52,7 +52,7 @@ static gboolean stream_eos (GMimeStream *stream);
 static int stream_reset (GMimeStream *stream);
 static gint64 stream_seek (GMimeStream *stream, gint64 offset, GMimeSeekWhence whence);
 static gint64 stream_tell (GMimeStream *stream);
-static ssize_t stream_length (GMimeStream *stream);
+static gint64 stream_length (GMimeStream *stream);
 static GMimeStream *stream_substream (GMimeStream *stream, gint64 start, gint64 end);
 
 
@@ -527,11 +527,11 @@ stream_tell (GMimeStream *stream)
 	return stream->position;
 }
 
-static ssize_t
+static gint64
 stream_length (GMimeStream *stream)
 {
 	GMimeStreamCat *cat = GMIME_STREAM_CAT (stream);
-	ssize_t len, total = 0;
+	gint64 len, total = 0;
 	struct _cat_node *n;
 	
 	if (stream->bound_end != -1)
@@ -563,7 +563,7 @@ stream_substream (GMimeStream *stream, gint64 start, gint64 end)
 	gint64 offset = 0, subend = 0;
 	GMimeStream *substream;
 	struct _cat_node *n;
-	ssize_t len;
+	gint64 len;
 	
 	d(fprintf (stderr, "GMimeStreamCat::substream (%p, %ld, %ld)\n", stream, start, end));
